Fixed a couple of things that didn't get done in the PR
Fixed another instance of snake case variables

Went with $skuPattern instead of $skuWildcard - since the input
is a pattern and not necessarily *just* a wildcard

I moved the error message construction to inside of the conditional
branch where we know an error occured.  It feels a bit odd to
me to begin constructing the error message before there's been
a problem.

Defaulted the $skuPattern to null - it shouldn't be a required parameter
for the loadRandomProduct method

Simplifed the conditional for getShippingMethodCode - I like to keep
conditionals dumbed down as much as possible.